Keep track of all Display.setCurrent() to ensure Alert works when switching display
We have always had problems, if we call setDisplay in the same function as creating
an Alert. Once the alert has been dismissed, the old Screen could still be shown,
rather than the new own. This was caused by the asynchronous nature of the setDisplay()
call, which might not have executed, when the Alert checks which screen to display
after it is dismissed.

The solution taken here is to manually keep track of all setDisplays and explicitly
pass this in when creating an Alert.