BugFix: Threading problems in FsDiscover
The FsDiscover was not showing the filesystem roots reliably.

It turns out, this was a threading issue. The thread reading
the filsystem roots was created before the init routine was
finished. The inti routines could overwrite the correctly
created list in the helper thread with an empty list from init.