Refactoring: Prepare for customizable ways
Refactor the code, such that style information is no longer
stored as a bunch of constants, but in a datastructure similar
to that of POIs.

This is a preparation for the fully customizable style-file.xml.
At the moment some of the way types are hardcoded, as there is no
support from the data format yet.

This patch reduces features (not all way types were converted)