BugFix: Numeric imprecision caused target way not to be found in Way.paint()
Rather than convert the truncated 16bit positions of the ways into floats and compare
it with the non truncated floats of the target, which will most of the times not
equal, truncate the target floats into shorts. Furthermore, don't test for equals,
but for close to equals.

We already did most of this in the SingleTile.walk code.  We should probably
get rid of the code duplication here anyway.