Another attempt to get runtime detection of jsr-179 working
Haveing to recompile the midlet for each combination of jsr's is tedious,
so a runtime detection is nicer. My previous attempt didn't seem to work very
well. The current solution is based on a tips and tricks from the sony ericsson
webpage (http://developer.sonyericsson.com/thread/44652).

Seperate out the implementation from the interface. Make sure that non of the interface
methods contain any reference to the jsr and then load the implementation through
the Class.forName(). With this, there should be no link from the main application
to the jsr other than through a text string.