Optimisation: Cache data a bit more aggressivly
So far we have deleted tiles and names from the caches if they haven't been used for a
fixed number of iterations. Instead we now check how much free memory we have and only
free up the caches if there is memory pressure. As we hold more data in memory, this may
speed up some operations such as repeated searches or zooming in and out.

As long as we are in a single tasked environment, there shouldn't be much negative impact.
In multi-tasked environments such as the Nokia S60 or newer Sony Ericsson phones, this might
have detrimental effects on other applications running on the phone, so this option needs
to be configurable in future.