Usability: Rename the back button to cancel in the configurartion dialog
This addresses a small part of #1942823. As the Ok button is often hidden
in the "more" options menu, it is not particularly intuitive that the back
butten reverts all the changes. By renaming it into a cancel button this
becomes a bit more obvious. The fact that the ok butten is hidden unfortunately
remains.