BugFix: calling setTitle within the paint method may cause infinite loops
According to the j2me spec, setTitle should cause a repaint to be issued.
So if you call setTitle in the paint method as we did, this leads to endless
loops of repainting. This is a big resource hog and can more or less starve
lower priority threads such as the ImageCollector thread. It seems as if
this bug only effected some devices (Motorola), as perhaps repeated
setting of the title to null did not cause repaints on all implementations