MicroOptimisation: Introduce a forward projection that takes relative coordinates
As all of our coordinates in the single tiles are in relative coordinates, but the
projection fuction took absolute coordinates, we always had to convert. By providing
a projection function that deals with relative coordinates directly we can save some
conversions. As this is on the fast path, this hopefully saves a small but measurable
difference.

This doesn't actually use the new methods yet