[ 1981228 ] Display area names in map
This patch displays the area names (and if zoomed in - the descriptions) in
the map:

It determines the centre of the area to determine the position for the text
(that's not perfect if you have i.e. an L-formed area but looks ok for most
cases)

The implementation also determines how many chars of the name will fit into
the area and just draw as many as will fit. If the name does not fit
completely it will add "..." to it.

If you zoom in to a scale<15000 then it will also display the way
description (and the name if available, i.e. "park" or "Olympiapark
(park)".

However - one problem with the patch is I can not figure out how to draw
the names on top of the ways. Hopefully I've prepared this by putting the
name drawing into a dedicated function that is currently just called from
paintAsArea().

Bumps Map version to 15