Rearranged some code to prevent some crashes when ImageCollector is not loaded
When the ImageCollector is not correctly initialised, such as if the wrong map-version
is used, then there were a bunch of null pointer exceptions being thrown. With this patch,
we try and at least not crash and potentially still be able to record gps data. Obviously,
it won't display a map in that case, so it is a bit useless.