Usability: Allow to search and display a street without setting it as a target
This allows to search for a road or POI and display it on the map, without
setting it as a target and thus leaving the current target unaffected.

This can improve usability when not having a GPS. For example if one
wants to navigate to A, then one can search for A and select it as a target.
One then has to navigate back to where one currently is. If you don't have
GPS. you had to manually move the curser to the current location. Now
one search for the streatname to position the cursor on the current location