Add an autoreconnect feature to the bluetooth connection
There seem to be some bluetooth devices that drop the connection after a while.
With this patch, if a broken connection is detected it will try and reconnect to
the receiver for a while and then give up if it didn't succeed. If it does
succeed, the Gps reception and logging should continue as normal.

This option is configurable in the Location Receiver setup menu.

This commit also contains the rest of the previouse commit that was
incorrectly forgotten during the commit.