Feature: Add fulltext search
This patch allows to perform a fulltext search. I.e. to enter a search string, which
then gets tested to see which names contain it. This allows to get around the fact
that one sometimes does not know the beginning of the name. For example does the
pub name start with "the", or is there a prefix such as "west" before the name.

However, there are several shortcommings to this code, so it is currently still
disabled. To enable it e.g. for testing, add back in the commented out line in GuiSearch.

Three major shortcommings:

1) The solution is very ugly. The way it works is if you enter a search string,
it goes through every name in the name list. This is not ideal, but probably not avoidable.
Much worse is, that we currently have no way to map full names to map features. So to get
arround this, each search result from the linear search then gets turned into the keynumbers
that one would have to press to get this name. That number string is then used to harness
the current search code. As each results has to go through the full search, retrieving
many names gets very slow. Overall the performance however seems surprisingly good for the
way it is doing it

2) It dupplicates the code in Osm2GpsMid to cannonicalise the name to key numbers

3) Not thread safe: The code to do the search currently spanns accross at least 3
 different threads. I don't think I have added all the necessary synchronisation
primatives into the code to make this work race free

4) Internationalisation: As it goes throught the cannonicalisation step, this probably
breaks for some of the non english chararcters

If anyone has better solutions please let me know. I felt it hopefully shouldn't harm
to commit this interime solution for the moment though and hopefully improve it later
before activating it.