Reorganised the Main menu in the Map screen
The menu in the main map screen was getting too long, especially
With yet another menu option for the sound recording.

This restructures the menu into a two layer menu, with the
Recordings... and Routing... option each opening a sub menu
that contains the relevant options.

I hope other people agree with this and find it more useable than
the one dimensional menu