BugFix: If the keylock was aktivated auto route-recalculation could get stuck in a loop
The route recalculation was triggered by calling a commandAction(). If the keyboard lock
was active, this however only popped up an error message.  It could also get into a loop.

As all the internal usage of commandAction does not use a Displayable, only prevent
key access if the displayable parameter was non null.