BugFix: NullPointerException in showRouting when no route was calculated yet
If a target was selected and now route calculated, then we would get
null pointer exceptions. Move that part into a if (route != null) check.

There still seem to be more null pointer exceptions happening in showRouting(),
but are hard to trigger. Could this be during route re-calculation due to different
threads updating and reading from the route array? i.e. a missing synchronisation?