Patch: [ 2078078 ] Adds GUI to enter a waypoint with coordinates
This is a patch by Markus Bauerle with one minor modification to the constructor of WayPoint.

Adds GUI to enter a waypoint with coordinates

It may be necessary for the user, while he is out in the field, to create a
waypoint by typing in coordinates. Especially Geocachers needs this often.

This patch adds the necessary GUI under the new "Recordings..." menu just
below "Add waypoint". (I took the liberty to reduce "Add a waypoint" to
"Add waypoint" in this patch as I think this is easier to read.)

The gimmick of this GUI is that you can enter decimal places in any field.
So you can either specify 48 30' 45" or 48 30.75' or 48.5125 and
leave the other fields empty. It will all be put together correctly.
It also checks if the values are in range to prevent the creation of
waypoints that are outside the coordinate system