BugFix: Map screen was not always correctly redrawn on startup
There was a race condition between the needsRedraw, the wait and the notify
in the ImageCollector thread loop. This could cause that the notify was sent
before the thread wait()ed. It mainly happened when the ImageCollector was created
at the beginning of the midlet, or when the size of the canvas changed

Now we simply have an extra check and only wait if needRedraw is false.
This presumably still leaves a small oportunity for a race condition, as it is not all
in a synchronized block, but hopefully it should be sufficiently unlikely now.