BugFix: Switching to full screen caused the projection to be off center
The image collector used to have a margin around the borders to ensure
smoother scrolling. I.e. the backbuffer was larger than the screen so that
even when the map center slighlty moved, the screen was still within the
backbuffer size and so a complete screen was shown while the imageCollector
thread rerendered the new backbuffer in the background. Given that the exact
size of the backbuffer didn't matter, we only updated the size of the backbuffer
when the canvas size increased, such as switching from normal to full screen.

Since some time, this assumption is no longer valid and the code instead assumes
the PaintContext to be exactly the screen size. Therefore we need to update
the size of the backbuffer on every size change rather than just when the size
increases.