Extend "Search/POI Overview" to "Search=>Overview/Filter Map"
"Search=>Overview/Filter Map" or double-press "*"
shows the chosen hideable POI/Area/Way types
in the map regardless of zoom level for an overview.

This can be useful to get a quick overview of a certain POI/Area/Way type in the area, i.e. hotels or railways or "restaurants AND pubs"

You can optionally choose to show POI/Area/Way types that are
 not selected in the Overview list normal or filter them out from display.