parse information if way type is routable from legend and put this into the WayDescription.
This is a preparation to be able to quickly distinguish between routable and non-routable ways. That's probably necessary as tramways and cycleways might be along the road, so if GPS or map is not exact enough to be on the proper way, we would try to start car routing from a tramway or a river, which would not work.