In addition to actualWay (which is displayed in the status bar) also determine the nearest routable way (according to wayDesc) which is closest to the center of the map.
That's necessary as tramways and cycleways might be along the road, so if GPS or map is not exact enough to be on the proper way, we would try to start car routing from a tramway or a river, which would not work.

So with these changes we'll always start route calculation from the nearest routable way according to wayDesc.