Quote from [ gpsmid-Patches-2461493 ] Multiple selection of tracks (mbaeurle):
Until now, you had to delete every recorded track one by one.
This was the main reason for me to implement the changes found in this patch.
Back when I started, you could also only export the tracks one by one.

I have implemented, much like the GUI found for waypoints, the possibility to select multiple tracks. There are also the menu entries "Select all" and "Deselect all".
The actions "Export" and "Delete" now work on the selected tracks.
"Display" and "Rename" use the first selected track for now. I'm not sure if it makes sense to change them to also work on the whole selection, especially displaying could run into trouble if a lot of long tracks are chosen. I wanted to get this patch finished so I left them mostly untouched for now. I renamed them to "Display first sel." and "Rename first sel." to make it clear that they use only the first track. I don't like the long description with the abbreviation very much but I had no better idea.
I removed "Export all" to keep the menu as short as possible. The action can still be done easily with a "Select all" before choosing "Export".

During export, the user is informed about the progress as another line reading "Exporting 2008-12...." is added to the alert box when the exporting of the next track is started.
Delete also gives feedback by writing "Deleting X tracks."
Both are finished by the message "Finished!" in the alert box. The alert box can't be dismissed until the action is finished to make sure the user does not interfere with the action.

There are two things that are not perfect at the moment, at least on my SE W800:
1) The alert box is repositioned to the beginning of the text after every addition of text. So you won't see the current progress and the message "Finished!" unless you scroll down yourself.
2) Depending on which tracks you delete, the alert box is not opened until the whole action is finished, so in these case, avoiding interference of the user with the action does not work. But the same is also true for the current GUI which never shows any feedback. To overcome this, the delete also has to run in a thread of its own.
This is increased by the fact that it takes very long to delete the last tracks in the list while leaving the first ones. It seems that the class library is doing a lot of moving of data here.
I'd like to hear feedback from other users how their phones behave here.

One thing I also changed on the side: The alert box when exporting waypoints has no timeout any more. It always disturbed me very much to have no feedback about the action if it took a little longer and I did not watch the phone all the time. If you didn't look during the 5 s after it finished, you didn't know if it was successful or not!

I just updated from CVS, so this patch is based on the very latest headrevision (Dec 23, ca. 14:00). The new GUIosmWayDisplay is taken into account and needed, among others, a change as I had to change the UploadListener interface to be able to give progress updates to the user.
As there are many changes to GuiGpx.java and Gpx.java, I have uploaded my versions of the affected files as well.

---

Note from sk750:
"...first sel."
renamed to
"...Sel 1"