Explicitly mark the commandAction function abstract and remove the implements CommandListener from Trace
On some phones like the k750, the jvm was not finding the commandAction method
when calling it from the KeyCommandCanvas. By explicitly makeing it abstract
and removing the implements CommandListener from Trace, the jvm should have an
easier job of finding the correct commandAction method.

Hopefully this solves the problem.