BugFix/Refactoring: Refactor routine to determin utf-8 encoding string into the Configuration class
There were several places where we tried to determin what the correct charecter encoding string for utf-8 is.
Refactor these to use the same code. This also fixes a bug in loading the charMap in NumberCanon, where we
didn't test for different encoding strings at all, causing it to fail on some phones