[Feature] Initial attempts in the direction of supporting content handlers
With content handlers (JSR211) it is possible to register a program to handle certain types
of data. For example, we can register a link, so that when a different Program has some coordinates to display,
it can call the content handler and GpsMid is invoked to display the map to show these coordinates.

This patch is based on code from Niko Berger found in Feature Request (2511260)

This is initial work and it is unlikely to work in its current form and is untested. Furthermore, it is guarded by the preprocessor
directive for JSR-211, which non of our default builds include, so in order to enable it you will need to change your custom-devices.xml