[ 2496655 ] Waypoints inside GPX tracklog, survey mode
This is a patch to write waypoints added along the way into the GPX
tracklog. When the gpx log is loaded into potlatch or JOSM the waypoints
are visible along the track and it's straightforward to add information to
OSM.

This is a slightly modified version of jkpl's patch to not store the waypoints
in-line of the actual track, but at the end. This is to not have to break the track apart