Feature: First primitive OpenStreetMap editing support for way tags
With this commit it is now possible to change individual tags on ways
in OpenStreetMap from within GpsMid, assuming one has Internet access on the phone.

One can select to download the raw data for a way from the OpenStreetMap server
and display the tags. These can then be added to, be removed or changed and reuploaded

This can be useful to add some tags such as e.g. maxspeed or name or other simple properties of a way.

It is currently not possible to change any nodes of the way, create new ways or delete ways. Furthermore,
changes made to the ways are not directly reflected in GpsMid and have to be imported through a new run of
Osm2GpsMid.