Patch [ 2541755 ] Avoid concurrent accesses to GpxTile
by Markus Baeurle

Currently, there is one tree structure of GpxTile instances to store
and display both trackpoints and waypoints.  While this suggests
itself at first, I think there are some disadvantages to this (not
necessarily in order of importance): 1) If the user is moving, track
points are added from the LocationProducer thread - typically with one
call to addTrkPt() every second. If he adds a waypoint, addWayPt() is
called in parallel from the UI thread. Right now, they are not
synchronized. Both add data to the same structure in any case.  2) We
have to assume there are some waypoints at startup which will be put
in a tile (or more if there are more than 90). So when track logging
is started, the track points will have to be added to this structure
with splitting decisions already made. They are not necessarily the
split decisions that are the best for the track points.  3) Generally,
the splitting decisions of both types may not fit. Maybe tile
splitting is started because trackpoints get added but the splitting
may not be ideal for the waypoints.

So I think it is better to split the tiles for trackpoints and
waypoints into two different classes. This both separates accesses to
these instances and splitting decisions within them. (I have also
discussed this with APM before changing it.)  Overhead should be small
as the duplicated code is also small. The tile instances themselves do
not need much memory. Most of the memory consumed is per trackpoint or
waypoint no matter where it is actually stored.  You may argue that it
is not strictly necessary to make this split and that it should be
possible to handle them both. I cannot 100% prove the opposite but I'm
sure it is easier to maintain this way and especially easier to be
sure there are no interferences between the two storing / displaying
actions this way.