Use the device framework to en/disable OpenStreetMap editing support in GpsMid
As the OpenStreetMap editing support adds quite a bit of code and additional data
in the map format, but will probably only be used by a small portion of our users,
as it requires internet access on the phone, we should hide it behind a preprocessor directive.
However, in order not to have to recompile GpsMid, I would like to include a version
of GpsMid that is enabled for editing in Osm2GpsMid. In order to do this, I have
(miss)used the custom-api framework. I have added a custom api "osm-editing" and used
a preprocessor to hide all the editing code behind this api. I also defined a new
device Generic/editing, that is identical to Generic/full, but also includes osm-editing