patch [2601338]: Remove "Erase xxx bytes" message
While the message "Erase xxx bytes", which is output after the connection
to the GPS receiver was established, was probably very useful while
initially implementing the NMEA message parsing, it provides an information
to the end users for which they have no use. In contrary, it might confuse
users as they might wonder where these bytes were erased and why and if
this means anything is wrong.
So this patch removes this message and changes it to a debug message
instead.
Plus it does a lot of code cleanup, mainly fixing typos and adding white
spaces.

Contains some additional code formatting in addition to the original patch