[ gpsmid-Patches-2619641 ] Another optimization for drawing waypoint labels by mbaeurle:
It is not necessary to copy every waypoint name to the StringBuffer, the length can be checked directly in the String class.
So this saves one copy to the StringBuffer and one toString() for every waypoint whose name is shorter than 25 characters.
Also, maxLen is just the copy of a constant and the StringBuffer is emptied inside the for-loop, so both don't need to be initialised every time inside the for-loop.